Refactored and Reorganized Create/Edit Gradeables Instructor Interface (#2068)
* Twig setup and home page template

* Barebones UI with navigation

* Initial Create page and team inheritance routing

* Fixed "Required" text not disappearing on radio buttons

* Simplified action conditionals and php variables

* Removed Max Team Count from Create page

* Added checkForm method for Create page

* Improved indentation of Create template

* Added Basic Graders template

* Finished UI for Graders page

This contains basic dynamic features, but does not have 'checkForm' validation yet.

* Fixed typo (add->new)

* Better switching based on gradeable type

* Better switching based on team-ness

* Added Dates Page

This includes both the check method and all inputs with the old date picker

* Improved Dates page layout.

Aligned all of the date boxes

* Added PDF page settings to Graders page

* Fixed Rotating gradeable history not displaying

* Added Auto-grading and Student settings pages

This content may be better off somewhere else, but now its in twig with appropriate validation and events.

* Added Syllabus Bucket input

* Added Rubric Page with Checkpoints/Numeric Entry

These are almost directly copied from the old interface.

* Improved formatting on rotating grader history

* Reverted database changes

inheritance will be added in a future PR.

* Electronic rubric and Moved misc items

* Improved electronic rubric view

* Fixed escaping json

* Fixed Grader group numbering

* Autograder config loads in properly

* Misc UI improvements

* Fixed buggy "Points" column in electronic rubric

* Rotating Data history should not appear if no history exists

* Page doesn't display until form loaded

This prevents the entire page from displaying before the DOM is ready and the navigation is loaded.

* Fixed rotating grader submission

* Basic navigation and autonavigation on 'create'

* Fixed typos

* Relocating settings and misc UI/nav fixes

* Misc UI improvements

* Nav bar style and misc fixes

* Updated Styling and moved "use repo" to 'create'

* Navigation redirection, config path checking, and id checking

fixed navigation redirection so it navigates to the correct "next" page.  The config path is no longer being checked (good default).  The gradeable ID is being properly checked too.

* fixed inheritance on checkpoints and numeric, electronic still broken

* Few navigation behavior fixes and form checking

* Fixed rubric inheritance for add_template

When making a new gradeable from a template, the rubric does not get inherited for any mode and the rubric starts in the initial state as if the template was not used.

* Fixed merge mistake (auto navigation on 'create')

* Fixed most of nav bar problems

* Readded maximum team size input

* Manual Grading Date disapperas when manual grading off

* Moved around UI elements and cleaned up duplicate code

* Moved scary message and adjusted title element type

* misc cleanup

* Autograding config better defaults and error handling

* Fixed misc text, alignment, and late days editible

before late days were not loaded in when generated

* Fixed checkpoints/numeric rubic not loading in

* Updated labels and indicating why fields uneditable

* Improved comments and "warning" lets you cancel

This addresses #2076 parts 1 and 3.

* Removed outdated/misleading comments

This isn't behavior we would want moving forward, so the comment shouldn't be there.

* Fixed grader assignment visibility and loading

I mistakenly removed a piece of code from the grader data compilation.

* Improved width of dates boxes

These were not the correct width on some systems, but now they are a fixed with that is guaranteed to be larger than a valid date.

* Missed content from fixing rotating graders

part of 64f6897d1ae734fa611096b3a304e40df9bd124c

* Removed unnecessary loop

* Fixed undefined variable error

* Revert "Removed unnecessary loop"

This reverts commit 537b28482557d19e131b52334b497d8ce8d58e90.

* Rotating Graders checkboxes show up properly

* Moved Scary 'Grading Begun' Message to be on all rubrics.

* just wording tweak

* Manual grading open date appears for checkpoint/numeric