Split Manual/Auto graded gradeable and added Autograding Config model (#2288)
* Split GradedGradeable in two

* Updated DB Queries for new split and efficiency

The queries are now setup to handle either missing TA grading or missing auto grading.  If its missing both, then we have no way of knowing anything.  Also, the gradeable data query can query for multiple gradeables at once.

* Simplified GradedComponent and AutogradingVersion

Construction of these are less restrictive and more convenient

* Added Autograding Config model to new gradeable model

This model is a more explicit representation of the autograding config data that is typically used by the gradeable and its consumers instead of an array.

* Documentation fix

* Moved autograding percent calculation

Its now in the version class, so each version can have its percent calculated and the AutoGradedGradeable juts gets the active version percents. This also requires the version to have access to the graded gradeable, so that was readded to its constructor.

* Updated phpdoc property type

* Added getter for GradedGradeable in TaGradedGradeable

* Removed unused private setter from GradedComponent

Graded components don't need to know about the graded gradeable

* Updated PercentGraded calculation

WIth some new information about how peer grades are handled.

* Added SubmissionTextBox class and updated config comments

I completed the documentation for properties on the AutogradingConfig class and added  a simple SubmissionTextBox class to encapsulate the data in a Textbox submission.

* Fixed gradeable data selector

I forgot some grouping symbols here.

* Added per-submitter graded gradeable query

Also renamed the gradeable data retrieval method to be consistent

* Improved exception consistency and documentation

* Fixed user/team id retrieval

Now it should use the data retrieved from the user table and not be dependent on the user id from the gradeable data table

* Renamed visible to NonHidden for consistency

* Fully qualified Gradeable type

* Factored percentage calculation into its own method

This was being done in a few spots, so I made it its own method