New Gradeable model integration into Simple Grader interface (#2406)
* Split GradedGradeable in two

* Updated DB Queries for new split and efficiency

The queries are now setup to handle either missing TA grading or missing auto grading.  If its missing both, then we have no way of knowing anything.  Also, the gradeable data query can query for multiple gradeables at once.

* Simplified GradedComponent and AutogradingVersion

Construction of these are less restrictive and more convenient

* Added Autograding Config model to new gradeable model

This model is a more explicit representation of the autograding config data that is typically used by the gradeable and its consumers instead of an array.

* Documentation fix

* Moved autograding percent calculation

Its now in the version class, so each version can have its percent calculated and the AutoGradedGradeable juts gets the active version percents. This also requires the version to have access to the graded gradeable, so that was readded to its constructor.

* Updated phpdoc property type

* Added getter for GradedGradeable in TaGradedGradeable

* Removed unused private setter from GradedComponent

Graded components don't need to know about the graded gradeable

* Updated PercentGraded calculation

WIth some new information about how peer grades are handled.

* Added SubmissionTextBox class and updated config comments

I completed the documentation for properties on the AutogradingConfig class and added  a simple SubmissionTextBox class to encapsulate the data in a Textbox submission.

* Marks handled better in GradedComponent

This allows for keeping track of removed/added marks and allows the user to set mark ids instead of required mark objects.  Invalid ids are filtered out when setting and don't throw an error.

* Fixed GradedComponent score handling

* Improved phpdoc types of TaGradedGradealbe

* Simplified some syntax

* Added db methods to create/update ta grades

* Improved defaults for graded constructors

* Few style consistencies

* Fixed gradeable data selector

I forgot some grouping symbols here.

* Added per-submitter graded gradeable query

Also renamed the gradeable data retrieval method to be consistent

* Added all-purpose 'save' method for grades

Since it is possible to tell if the grade is new or not based on its id, the database calling methods handle whether to create/update instead of the user needing to make an identical check every time.

* Grade/Save/printlab simple grader controller

* Added per-submitter graded gradeable query

Also renamed the gradeable data retrieval method to be consistent

* Improved exception consistency and documentation

* Fixed user/team id retrieval

Now it should use the data retrieved from the user table and not be dependent on the user id from the gradeable data table

* Renamed visible to NonHidden for consistency

* Fully qualified Gradeable type

* Factored percentage calculation into its own method

This was being done in a few spots, so I made it its own method

* UploadCSV

* Clarified return types on db methods

* Added param to TaGradedGradeable getter to generate

* Fixed last exception comment

* Added method to get graded components easily

This method will get/generate a component for a ta graded gradeable based on the supplied parameters.  It kind of just does what you want in each case, but thta should be documented more clearly

* Initial work on primary view

* Improved documentation for how arrays vs values are handled

* Fixed spacing / style

* Comprehensive documentation for getGradedComponent

While this function does what youd want it to do in any situation, my interpretation of what the best option is is not what everyone might think, so I wrote down exactly how this function behave.

* More consistent wording

* Improved naming for get/create methods

* Removed unnecessary code

* Misc style and method name fixes

* Fixed parameter type enforcement in TaGradedGradeable

* Viewing works

* Fixed sort function parameters

* Fixed saving and numeric/text view

* Added defaults for null grades

This is to prevent NAN errors on the site

* Fixed spacing

* fixed method name

* Fixed null section label

* Fixed ordering

This was causing the total column to be incorrectly calculated

* Clarifying comment

* Updated comment

* Updated test for NULL section

This was being displayed as blank on master, but it looked like the intent was for NULL to show up in the twig template, so it should be displaying 'NULL' now and the test looks for that.