New Gradeable model integration into Nav page (#2407)
* Split GradedGradeable in two

* Updated DB Queries for new split and efficiency

The queries are now setup to handle either missing TA grading or missing auto grading.  If its missing both, then we have no way of knowing anything.  Also, the gradeable data query can query for multiple gradeables at once.

* Simplified GradedComponent and AutogradingVersion

Construction of these are less restrictive and more convenient

* Added Autograding Config model to new gradeable model

This model is a more explicit representation of the autograding config data that is typically used by the gradeable and its consumers instead of an array.

* Gradeable lazy loads teams

In addition to knowing if there are any teams, it is often useful to know what those teams are.  Now, these teams are lazy loaded into the gradeable instance (not serialized in toArray())

* GradeableList and GradeableListTester use new model

* Grade percent calculation for grader in User model

Instead of calculating the percent a grader has graded an assignment in the view, this is now done in the user model

* Nav View/Controller use new model

Optimized with the new model and db access methods, the new view uses fewer queries and is faster to load.

* Documentation fix

* Moved autograding percent calculation

Its now in the version class, so each version can have its percent calculated and the AutoGradedGradeable juts gets the active version percents. This also requires the version to have access to the graded gradeable, so that was readded to its constructor.

* Updated phpdoc property type

* Added getter for GradedGradeable in TaGradedGradeable

* Removed unused private setter from GradedComponent

Graded components don't need to know about the graded gradeable

* Updated PercentGraded calculation

WIth some new information about how peer grades are handled.

* Added SubmissionTextBox class and updated config comments

I completed the documentation for properties on the AutogradingConfig class and added  a simple SubmissionTextBox class to encapsulate the data in a Textbox submission.

* Duplicated GradeableList

Since the GradeableList gets used in other places, the slightly modified version using the new model is now in the gradeable folder.

* Updated use statement

* Fixed gradeable data selector

I forgot some grouping symbols here.

* GradedGradeable query gets data for all teams/users requested

Now all teams / users requested will get data returned regardless of if any grade data exist.

* Added gradeable method for checking build error state

* Updated name from merge

* Updated usage of getGradedGradeables

* NavigationView tolerant to missing GradedGradeables

* Small comment update

* Fixed gradeable anySubmissions method

This meant that gradeables that shouldn't be deletable were able to be deleted.

* comment

* Very annoying...

* Moved grading progress to gradeable

* Fixed TA percent comparison

* Fixed team submit button

* team submit button

* Fixed team button

* Finally fixed team submit button

* remove file

* improved spacing

* Removed hard-coded path

* Moved GradeableListTester