[Bugfix:InstructorUI] student search function overlap (#9180)
* 1

i think the original purpose of the function for student search is to have it follow when scroll. So the user may use the search function whenever instead of scrolls to top every time. But I removed all related codes since the purpose of issue #9177 is to have the student search function fixed at its position.

* Update simple-grading.js

1