[Feature:TAGrading] NavigationBar for grading_beta (#9483)
### Dependencies
This PR is a branch of PR #9454, and thus PR #9454 should be merged
first.

### What is the new behavior?
The grading_beta page now has the NavigationBar. The Home button takes
one back to the Details page. The Previous Student and Next Student
buttons do nothing currently. (That is for the next PR.) The full-screen
button works as expected. The settings button does nothing currently,
and the Progress Bar currently is fixed at 65%.

### What to test?
- The Home button on grading_beta does indeed take you to the
Gradeable's details page.

- The Full Screen button on grading_beta does indeed fullscreen the
page.

- If Anonymous Mode is enabled as an instructor on the Details page, the
grading_beta does not display the Student Information on the left side
of the navigation bar.
- If Anonymous Mode is disabled, the Student Informtaion on the left
side of the navigation bar is visible.

- If you copy the current URL of a grading_beta page to a browser that
is logged in as student, the Student Information on the left side of the
navigation bar should not appear.

### How do I access grading_beta?
1. On the main course's page, pick the Grade button for an assignment on
the Gradeable page


![image](https://github.com/Submitty/Submitty/assets/51209504/26c6e76e-3e30-46a8-8fa6-9f8cc515635a)

2. You are now on the Gradeable's Details page. Click the Grade/Score
button for a particular student, and if this Gradeable is an
Electronic/Rubric (Rubric is the new name we are changing to) gradeable,
the TA Grading page will appear.


![image](https://github.com/Submitty/Submitty/assets/51209504/e80de0b0-37ad-44ab-a820-e092793a0a5b)

3. The URL will look something like this:


SERVER/courses/TERM/COURSE/gradeable/GRADEABLE_NAME/grading/grade?who_id=ANONYMOUS_STUDENT_HASH&sort=id&direction=ASC

Change /grading/ in the URL to /grading_beta/ to open grading_beta.


![image](https://github.com/Submitty/Submitty/assets/51209504/30869a69-8b5c-4ff9-a49a-38f37d89ad58)


4. grading_beta will look like the following:


![image](https://github.com/Submitty/Submitty/assets/51209504/49cea180-629d-4a6a-8c44-3a2f948b9efe)

---------

Co-authored-by: Mahi Pasarkar <47994802+mahi-pas@users.noreply.github.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>