[Feature:Developer] Integrate Cypress Cloud  (#9517)
### What is the current behavior?
Currently we just use the testing component of cypress.

### What is the new behavior?
This will add the setup to have cypress link with cypress cloud, which
will allow for easier parallelization, and it shows recordings of the
tests in the cloud.

### Other information?
~~The open source license allows for 100,000 runs (specific it('',
()=>{}) functions, not spec files) per month, so this is the testing
period to see if this is enough, if not, it may be better to manually
parallelize the tests and use cloud as a different feature, to see why
cypress is failing in CI and not locally, but we will see. I have also
communicated with some people at cypress, and they indicated that we may
be able to get more runs, but I haven't communicated more about that.~~

This information is old. We now have an OSS cypress cloud
'organization', which allows for unlimited runs, not 100,000/month.