[Bugfix:Developer] Fix usage of @prop (#9635)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Class properties in the PHP codebase are currently annotated with
docstring comments beginning with `@prop @var`. This is not valid phpdoc
syntax and results in the types being ignored by static analysis.
Additionally, the phrase `@property` was used interchangeably with
`@prop`. This is problematic as `@property` is a recognized keyword used
for defining virtual properties on classes.

### What is the new behavior?
All usage of `@prop` and its derivatives have been separated from `@var`
so as to not hinder static analysis. Incorrect uses of `@property` have
been replaced with `@prop`.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

The `@prop` keyword is a custom keyword used in PHP typedoc comments
throughout the Submitty codebase. Our code relies on it for certain
tasks such as the hydration of models. We also have derivatives such as
`@prop-read` and `@prop-write`, which serve as makeshift access controls
for properties.

It is worth discussing if there is a better approach to solving this
issue, and in which case whether or not this PR should be merged in as a
temporary solution while a more permanent refactor is under the process
of being worked on.