[SECURITY][Bugfix:System] Referrer Policy Update (#9772)
### What is the current behavior?
Before this patch, no referrer policy is defined, and the browser's
default behavior regarding the transmission of referrer information is
in effect. This may lead to the browser potentially sending the complete
URL as the referrer when navigating to different origins, which can have
privacy and security implications.

### What is the new behavior?
After applying this patch, the HTTP response sets the referrer policy to
"strict-origin-when-cross-origin." This ensures that the browser will
send the full URL as the referrer only when navigating to a cross-origin
resource, enhancing privacy and security by reducing unnecessary
disclosure of information to external websites.