[Refactor:System] Update Node Repo (#9756)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
There is currently a deprecation in the node installation script. This
currently halts installation for a minute and this method will likely be
removed in the future.

### What is the new behavior?
Rather than fetching the nodesource script, the debian repo is inserted
manually. This is the new suggested way to install node. This PR also
bumps Node 16 to 20.

### Other information?
Testing vagrant up here:
https://github.com/Submitty/Submitty/actions/runs/6149685803