[Testing:Developer] Add slack message on vagrant up failure (#9781)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently the vagrant up github action is a cron job, so it is run on a
schedule. Due to this, it only sends notifications to the last person to
edit the workflow file. If this person is away for any significant
period, the failing job can go unnoticed by other developers. This is a
problem, as new developers won't be able to reliably run 'vagrant up',
which will slow development.

### What is the new behavior?
This adds a notification VIA a slack app that notifies all who are in
the slack channel of the vagrant up failure.

![image](https://github.com/Submitty/Submitty/assets/46759635/21316cc6-9f04-4e77-b31d-4e4e7b232b88)

The first two messages were debugging, and the last one was a mock
failure to show it works.

Closes #9596