[Bugfix:TAGrading] Fix Incorrect Stats for Grader Scores (#9774)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Currently, in the grading stats page. the statistics available for each
grader (count,average,stddev) are not correctly calculated:
 

![image](https://github.com/Submitty/Submitty/assets/131885371/200de6f4-24c3-4d82-886e-f2b43707df17)
Notice how the stats for the grader for a specific component differ from
the overall stats per components.
Additionally, the stats for the grader per component don't interact with
the newly added filters.

### What is the new behavior?

![image](https://github.com/Submitty/Submitty/assets/131885371/dec2e5f9-5954-455d-8091-efa34986dcee)
The stats for the grader per component are correctly calculated and
interact with the filters
### Other information?
Additionally, this PR reformats some queries related to the statistics
for better code practise.