[Refactor:InstructorUI] Refactor queries to collect data earlier (#9806)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
The instructor gradeables page makes 108 queries each page load. A lot
of these queries are repetitive and can be refactored to use
significantly less queries.

### What is the new behavior?
The gradeables page now uses 82 queries per page load which is about a
20% improvement. The gradeable canDelete function can be pretty
expensive so I made it collect some data earlier and only query team
info if it is a team gradeable.

### Other information?
There is still more work to be done here. The most expensive thing on
the page is loading all of the status bars for grading but it uses a lot
of query and PHP logic.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>