[Bugfix:Developer] Fix vagrant up slack message (#9838)
### What is the current behavior?
After I was testing the slack message on failure, I accidentally deleted
the line to get the JOB ID on failure. This causes the vagrant up
notification to not have the JOB ID, so the link is bad.

### What is the new behavior?
The line to run this on failure has been re-added, this will make the
link accurate.