[Feature:Developer] Add notifications for good first issue (#9917)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Related to #9903 
Currently when good first issues are posted to the Github, there is no
notification to developers that an issue has been posted. If it is not
in your workflow to check issues regularly, they can be lost in the
hundreds of issues we currently have.

### What is the new behavior?
A workflow has been added to notify the Zulip channel on the labelling
of an issue as a 'good first issue'. The wording of the content isn't
concrete, so if there is something that should be added/removed it is
easily changeable.



![image](https://github.com/Submitty/Submitty/assets/46759635/cc9a7abb-4791-4d07-bcad-3652a900e088)