[Refactor:Submission] Luxon JS Date Library (#9909)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
We use the builtin JS date library in a number of places. While this
mostly works, we keep getting bit by browser implementation specific
stuff and pitfalls because it's not a very developer friendly interface.
This is especially a bit hairy as we then do bits of math with these
dates, which is equally error prone.
Related to #9731, however, do NOT close that issue because there are
more files that needs to be refactored.

### What is the new behavior?
Instead, we are using Luxon which abstracts this stuff (as well as takes
advantage of the Intl library for better international support). This
would also provide the same interface across all browsers so that we're
confident we're not going to hit a bug where the date displays properly
on Chrome, but not on Safari or FireFox.

---------

Co-authored-by: Youssef-Chip <131885371+Youssef-Chip@users.noreply.github.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>