[Feature:Calendar] Clickable buttons darken when hovering. (#9837)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Closes #9705. Calendar entries with links embedded appear no different
from those without. It is difficult to tell whether a button has a link
embedded without clicking it.

### What is the new behavior?
When hovering the mouse over a button with a link, the button becomes
slightly darker. Buttons without links do not change color. The
screenshot doesn't show the cursor, but it's hovering over the "Docker
Network Python" button in the top screenshot, but not in the bottom.

![clickcaldiff](https://github.com/Submitty/Submitty/assets/89590361/e5a96347-2516-4e13-be6a-69eb6aff2eeb)

![clickcallbef](https://github.com/Submitty/Submitty/assets/89590361/fea5f572-a8a8-44f9-b512-b896a8f6e523)

### Other information?
I built this on top of my previous branch, PR
https://github.com/Submitty/Submitty/pull/9810 . All the changes to CSS
files are related to that. This project only required editing
calendar.js

Tested by accessing calendar page and ensuring proper behavior of
entries with and without links.